requirements.txt
pygame==2.1.2
Pyganim==0.9.2
pyperclip==1.8.2
python-i18n==0.3.9
typing_extensions==4.3.0
